moved some header files into appropriate include folder
