+++ .gitignore
@@ -1,3 +1,4 @@
 *.user
 *.kdev4
 *.autosave
+*.vscode
