+++ include/cartesio_planning/goal/NSPG.h
@@ -1,4 +1,4 @@
-#include <ik/position_ik_solver.h>
+#include <cartesio_planning/ik/position_ik_solver.h>
 #include <chrono>
 #include <functional>
 #include <stdio.h>
