+++ src/nodes/goal_generation.h
@@ -4,7 +4,7 @@
 #include <XBotInterface/ModelInterface.h>
 #include <ros/node_handle.h>
 #include <cartesio_planning/validity_checker/validity_checker_context.h>
-#include <goal/goal_sampler.h>
+#include <cartesio_planning/goal/goal_sampler.h>
 #include <cartesian_interface/ros/RosServerClass.h>
 #include <cartesio_planning/CartesioGoal.h>
