+++ src/nodes/planner_executor.h
@@ -12,7 +12,7 @@
 
 #include <cartesian_interface/utils/RobotStatePublisher.h>
 
-#include "constraints/cartesian_constraint.h"
+#include <cartesio_planning/constraints/cartesian_constraint.h>
 #include <cartesio_planning/planner/cartesio_ompl_planner.h>
 #include <cartesio_planning/utils/robot_viz.h>
 #include <cartesio_planning/validity_checker/validity_predicate_aggregate.h>
