+++ src/planner/cartesian_trajectory_interpolation.cpp
@@ -1,5 +1,5 @@
 #include <cartesio_planning/interpolator/cartesian_trajectory_interpolation.h>
-#include "planner/trajectory_interpolation.h"
+#include <cartesio_planning/planner/trajectory_interpolation.h>
 #include <RobotInterfaceROS/ConfigFromParam.h>
 
 CartesianTrajectoryInterpolation::CartesianTrajectoryInterpolation()
